Set up generic AprilBot
This boilerplate creates some basic AprilBot functionality for
connecting to slack and handling some basic types of interactions.

I had to create types for the event handlers for now, though I opened an
issue in the Slack sdk in the hopes they might add types [1] to their
events.

This also swaps out the class-methods-use-this lint rule for a smarter
variety, which is probably something that the upstream maintainer should
consider adding as well [2].

[1] https://github.com/slackapi/node-slack-sdk/issues/1768
[2] https://github.com/iamturns/eslint-config-airbnb-typescript/issues/344